services/feature_computer_1m/requirements.txt
boto3==1.35.0
psycopg2-binary==2.9.9
numpy==1.26.4
pandas==2.2.0
